Add interface imports in webapp
