+++ webapp/src/components/Context/UserContext.tsx
@@ -2,7 +2,8 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react'
 
 import { api } from 'api'
 import { useLocalStorage } from 'utils/persist'
-import { LocalKey } from '@constants'
+import { LocalKey } from 'constants/LocalKey'
+import { User } from 'common-ts/dist/webServiceApi/interface/user'
 
 type ContextType = {
   user?: User
