+++ webapp/src/components/History/PaymentEntry.tsx
@@ -3,6 +3,7 @@ import VisibilityIcon from '@mui/icons-material/Visibility'
 import { TableCell, TableRow, Tooltip, Typography } from '@mui/material'
 import { useTranslation } from 'react-i18next'
 
+import { PaymentDetails } from 'common-ts/dist/webServiceApi/interface/payment'
 import { PaymentStatusCell } from './PaymentStatusCell'
 import { TokenPopup } from './TokenPopup'
