+++ webapp/src/components/Layouts/AdminLayout.tsx
@@ -4,8 +4,8 @@ import { Navigate } from 'react-router-dom'
 
 import routesBuilder from 'routesBuilder'
 import { UserContext } from 'components/Context/UserContext'
-import { Role } from '@constants'
 import { PanelLayout } from './PanelLayout'
+import { Role } from 'common-ts/dist/webServiceApi/interface/user'
 
 export const AdminLayout: React.FC = () => {
   const { user, isLoggedIn, loading } = useContext(UserContext)
