+++ webapp/src/components/Layouts/UserLayout.tsx
@@ -5,7 +5,7 @@ import { Navigate } from 'react-router-dom'
 import routesBuilder from 'routesBuilder'
 import { UserContext } from 'components/Context/UserContext'
 import { PanelLayout } from './PanelLayout'
-import { Role } from '@constants'
+import { Role } from 'common-ts/dist/webServiceApi/interface/user'
 
 export const UserLayout: React.FC = () => {
   const { isLoggedIn, loading, user } = useContext(UserContext)
