+++ webapp/src/components/QuickBuy/Stages/SetupStage.tsx
@@ -7,6 +7,7 @@ import { CardForm, CardFormButton } from 'components/Form/CardForm'
 import { StageProps } from 'components/QuickBuy/StageProps'
 import { TextInput } from 'components/Form/FormikInputs/TextInput'
 import { api } from 'api'
+import { PaymentForm } from 'common-ts/dist/webServiceApi/interface/payment'
 
 export const SetupStage: React.FC<StageProps> = ({ onNext, setPayment }) => {
   const { t } = useTranslation('quickBuy')
