+++ webapp/src/constants/LocalKey.ts
@@ -1,0 +1,6 @@
+export enum LocalKey {
+  TOKEN = 'token',
+  USER = 'user',
+  PAYMENT = 'paymentDetails',
+  TRANSACTION_TOKENS = 'transactionTokens'
+}
